Dialogue box added
nice dialogue box with typewriter effect
player can be now immobilized (for events such as talking to an npc, events cutscenes etc)

+++ Game/cmp_health.h
@@ -13,10 +13,10 @@ protected:
 public:
 	explicit HealthComponent(Entity *p);
 	HealthComponent() = delete;
-	int getHealth() { return health; }
-	int getMaxHealth() { return maxHealth; }
-	int getMaxMana() { return maxMana; }
-	int getMana() { return mana; }
+	float getHealth() { return health; }
+	float getMaxHealth() { return maxHealth; }
+	float getMaxMana() { return maxMana; }
+	float getMana() { return mana; }
 	void reduceHealth(float x);
 	void reset();
 	void reduceMana(float x);
